backend/requirements.txt
Django>=4.0.0
djangorestframework
django-cors-headers
channels
daphne
psycopg2-binary
boto3
redis
channels_redis
dj-database-url
djangorestframework-simplejwt
django-otp
qrcode
Pillow
pyotp